Improve font rendering
Correct HUD transparency, so that the HUD is now transparent and hence
can be drawn over the top of the rest of the game. Note that SDL2_TTF
currently ignores the alpha component of colors, which is probably a
bug (i.e. fonts will always be rendered at 255 alpha).

Also corrected possible segfault if HUD texture is not the same size as
the hud surface.